added vue router, reorganised routes
