+++ package.json
@@ -19,7 +19,8 @@
     "bulma": "^0.9.4",
     "neo4j-driver": "^4.4.7",
     "v-network-graph": "^0.6.5",
-    "vue": "^3.2.37"
+    "vue": "^3.2.37",
+    "vue-router": "^4.1.3"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^3.0.0",
