+++ src/App.vue
@@ -1,118 +1,27 @@
 <template>
-  <div id="graph">
-    <v-network-graph ref="graph" v-model:selected-nodes="graphData.selectedNodes"
-      v-model:selected-edges="graphData.selectedEdges" :nodes="graphData.nodes" :edges="graphData.edges"
-      :layouts="graphData.layouts" :configs="graphData.configs" :event-handlers="graphData.eventHandlers" />
-  </div>
-
-  <div id="config-modal" class="modal">
-    <div class="modal-background"></div>
-    <div class="modal-card">
-      <header class="modal-card-head">
-        <p class="modal-card-title">Configuration</p>
-        <button class="delete" aria-label="close"></button>
-      </header>
-      <section class="modal-card-body">
-        <form>
-          <label class="checkbox">
-            <input type="checkbox" id="d3-force-enabled" v-model="graphData.d3ForceEnabled" />
-            Auto-organise
-          </label>
-        </form>
-      </section>
-      <footer class="modal-card-foot">
-        <div class="container">
-          <button class="button is-primary">Done</button>
-        </div>
-      </footer>
-    </div>
-  </div>
-
-
   <Navbar></Navbar>
-  <div class="info-panel-outer">
-    <div class="info-panel-inner">
-      <div v-if="dataPanel.get('labels')" v-bind:id="dataPanel.get('id')" class="info-panel-inner-details-id">
-        <h3>Properties</h3>
-        <ul>
-          <li v-for="(label, index) in dataPanel.get('labels')" :key="index" contenteditable="true"
-            v-on:blur="onInputLabel">
-            {{ label }}</li>
-          <li v-if="dataPanel.get('objType') == 'Node'"><input type='button' value='Add Label'
-              onclick='document.addLabel()' />
-          </li>
-        </ul>
-        <table>
-          <tr>
-            <th>&lt;id&gt;</th>
-            <th>{{ dataPanel.get('id') }}</th>
-          </tr>
-          <!-- <p>{{dataPanel.properties}}</p> -->
-          <tr v-for="(value, property) in dataPanel.get('properties')" :key="value">
-            <th contenteditable="true" v-on:blur="onInputPropertyName">
-              {{ property }}
-            </th>
-            <th v-bind:id="'valueOf' + property" contenteditable="true" v-on:blur="onInputPropertyValue">
-              {{ value }}
-            </th>
-          </tr>
-          <tr><input type="button" value="Add Property" onclick="document.addProperty()" /></tr>
-        </table>
-      </div>
-      <div v-else-if="dataPanel.get('objType')">Creating {{ dataPanel.get('objType') }} in Neo4J... Hover over node
-        again later to
-        check
-      </div>
-      <div v-else>Hover over a node/edge to check it out</div>
-    </div>
-  </div>
-
+  <router-view/>
 </template>
 
 <script lang="ts">
 
 import { defineComponent, ref, reactive } from 'vue'
-import graphData from "./graphData"
+import graphData from './graphData';
 import Navbar from './components/Navbar.vue';
 
-var dataPanel = reactive(new Map<string, any>());
-dataPanel.set('labels', [])
-dataPanel.set('id', '123')
-dataPanel.set('properties', [])
-dataPanel.set('objType', [])
-// const graph = graphData.graph
-// const nodes = graphData.nodes
-// const edges = graphData.edges
-// const configs = graphData.configs
-// const layouts = graphData.layouts
-// const selectedNodes = graphData.selectedNodes
-// const selectedEdges = graphData.selectedEdges
-// const d3ForceEnabled = graphData.d3ForceEnabled
-// const eventHandlers = graphData.eventHandlers
 export default defineComponent({
   name: "App",
   components: {
-    Navbar
+    Navbar,
   },
   data() {
-    return {  dataPanel, 
-              graphData, 
-    }
+    return { graphData }
   },
   methods: {
     keyDown: function () {
       console.log("hello there")
     },
-    onInputPropertyName : function() {
 
-    },
-    onInputPropertyValue : function() {
-
-    },
-    onInputLabel : function (){
-
-    },
-   
     modalEvent: function () {
       // Functions to open and close a modal
       function openModal($el: any) {
