+++ src/main.ts
@@ -7,6 +7,7 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { faGithub } from "@fortawesome/free-brands-svg-icons";
 
 import App from "./App.vue"
+import router from "./router"
 import 'bulma/css/bulma.css'
 
 library.add(faGithub)
@@ -14,4 +15,5 @@ library.add(faGithub)
 const app = createApp(App)
 app.use(VNetworkGraph)
 app.component('font-awesome-icon', FontAwesomeIcon)
+app.use(router)
 app.mount("#app")
