+++ src/views/Explore.vue
@@ -1,0 +1,3 @@
+<template>
+dora d exploraaaa
+</template>
