+++ src/views/Home.vue
@@ -1,0 +1,3 @@
+<template>
+this is home truly
+</template>
